backend/pyproject.toml
[project]
name = "safetycare"
version = "1.0.0"
description = "Fall Detection System via IP Cam using MediaPipe and YOLOv8"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [
    { name = "SafetyCare Team" }
]
keywords = ["fall-detection", "computer-vision", "mediapipe", "yolov8", "fastapi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.125.0",
    "uvicorn[standard]>=0.40.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.21",
    "python-jose[cryptography]>=3.5.0",
    "bcrypt>=5.0.0",
    "httpx>=0.28.1",
    "websockets>=15.0.0",
    "opencv-python-headless>=4.12.0",
    "mediapipe>=0.10.21",
    "ultralytics>=8.3.0",
    "numpy>=2.2.0,<2.3.0",
    "pillow>=12.0.0",
    "structlog>=25.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.2.0",
    "ruff>=0.14.0",
    "mypy>=1.19.0",
    "pre-commit>=4.5.0",
]

[project.scripts]
safetycare = "safetycare.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/safetycare"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.2.0",
    "ruff>=0.14.0",
    "mypy>=1.19.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = ["E501", "B008", "B904"]

[tool.ruff.lint.isort]
known-first-party = ["safetycare"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "cv2.*",
    "mediapipe.*",
    "ultralytics.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=safetycare --cov-report=term-missing"
